Update lexer
In this modifications, we have added two methods to the lexer structure
The first one called GetCursorPosition is made to allows to get the current value of the lexer's attributes readPosition and position. the second one called SetPosition aims to set current values of the last mentioned lexer attribute.
the last one is called Input() which allows to get the value of the current input value.