Update parseIdentifier()
In this update, we modifiy the method parseIdentifier() in order to permit it to parse the identifier in the case of accessing to the member of the structure.

We also added the method parseAsExpression() to deal with SQL Select statement in the case of renaming the field in the clause select or in the clause from.