Update readNumber method in the lexer
Update readNumber method in the lexer.
we added "&& isDigit(l.peekChar())" before reading the float in order to be sure that the number we want to read is really a float. this modification is due to the fact that we notice when we read the first part of the range, this member is always a float even if, it's supposed to be integer.